Address review finding: drop redundant long cast in unknownService port assertion
Review finding:

This new assertion introduces a redundant (long) cast. .github/instructions/java-tests.instructions.md says: "Do not introduce redundant (long) casts in equalTo(longKey(...), value) when value is already an int - the equalTo(AttributeKey<Long>, int) overload exists." io.grpc.Server.getPort() returns int, and there is no ternary here forcing a boxed type, so equalTo(SERVER_PORT, server.getPort()) compiles through the int overload. The neighbouring RPC_GRPC_STATUS_CODE assertion in the same block genuinely needs its cast, because emitOldRpcSemconv() ? ... : null would otherwise box to Integer, but this one does not. The same rule was already applied to the three SERVER_PORT assertions this PR added to ArmeriaGrpcTest, which now read equalTo(SERVER_PORT, server.httpPort()) with no cast, so the two new tests are inconsistent with each other. Fix: drop the (long) cast, leaving equalTo(SERVER_PORT, server.getPort()).

Analysis: OpenTelemetryAssertions.equalTo is overloaded with equalTo(AttributeKey<Long>, int), which widens the int to long and delegates to the generic equalTo. Server.getPort() returns int and the argument is not part of a ternary that would box it to Integer, so removing the cast keeps the same overload resolution result and the same assertion behavior. The adjacent RPC_GRPC_STATUS_CODE assertion keeps its cast because emitOldRpcSemconv() ? Status.Code.UNIMPLEMENTED.value() : null would otherwise box to Integer and match no overload.

Upsides: The new assertion follows the repository test rule, and the two unknownService tests added by this pull request now write the port assertion the same way.

Downsides: No material downside identified.

+++ instrumentation/grpc-1.6/testing/src/main/java/io/opentelemetry/instrumentation/grpc/v1_6/AbstractGrpcTest.java
@@ -888,7 +888,7 @@ public abstract class AbstractGrpcTest {
                             RPC_RESPONSE_STATUS_CODE,
                             emitStableRpcSemconv() ? Status.Code.UNIMPLEMENTED.name() : null),
                         equalTo(SERVER_ADDRESS, "localhost"),
-                        equalTo(SERVER_PORT, (long) server.getPort()))));
+                        equalTo(SERVER_PORT, server.getPort()))));
 
     testing()
         .waitAndAssertTraces(
